requirements.txt
streamlit
gspread
google-auth
pandas
fpdf2
xlsxwriter
docxtpl